Add wireframe_testing crate and migrate tests
