+++ Cargo.toml
@@ -14,6 +14,7 @@ log = "0.4"
 
 [dev-dependencies]
 rstest = "0.18.2"
+wireframe_testing = { path = "./wireframe_testing" }
 
 [lints.clippy]
 pedantic = "warn"
