+++ wireframe_testing/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "wireframe_testing"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+tokio = { version = "1", features = ["macros", "rt", "io-util"] }
+wireframe = { path = ".." }
+bincode = "2"
+bytes = "1"
+
+[dev-dependencies]
+rstest = "0.18"
